backend/requirements.txt
# Backend Framework
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Database
pymongo==4.6.2
redis==5.0.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & ML
google-generativeai==0.3.2
numpy==1.26.4

# Speech Processing
gTTS==2.5.1
SpeechRecognition==3.10.1

# WebSocket
websockets==12.0

# PDF Generation
reportlab==4.1.0
matplotlib==3.8.3

# Utilities
pydantic==2.6.3
email-validator==2.1.1
python-dateutil==2.8.2